Añadido mensaje aviso inserción nueva clave API. Ahora cuando se inserta una clave API (haya ya una almacenada en localStorage o no), le pregunta al usuario si desea confirmar que quiere guardar una nueva clave de API (lo que haría perder la clave antigua guardada en localStorage en caso de haber alguna).

+++ my-chatgpt/chat.js
@@ -238,7 +238,7 @@ function actualizarConversacionUI() {
 function fEnviarButton(evento) {
     // console.log(promptbox.value);
     /*-- Descarta que no hay una clave API para poder interactuar con ChatGPT --*/
-    if (!clave_api) {
+    if (!clave_api && window.confirm("Estás a punto de guardar una nueva clave de la API. ¿Confirmar?")) {
         clave_api = promptbox.value
         localStorage.setItem("claveAPI", clave_api);
         mensajeEstado(false,
